requirements.txt
yt-dlp>=2023.12.30
openai-whisper>=20231117
google-generativeai>=0.3.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
trafilatura>=1.6.0

# 追加のキュレーション方法用ライブラリ
praw>=7.7.0  # Reddit API
tweepy>=4.14.0  # X（Twitter）API
beautifulsoup4>=4.12.0  # HTMLパース
feedparser>=6.0.10  # RSSフィード解析

